Move CourtCaseDetailsTabs and CourtCaseDetailsSingleTab to a new file
